Add trapceptions (#10)
This adds spike traps. Similarly to how monsters work, traps emit a
signal on colliding. There are however a few differences between
a monster and a trap. Traps have states (Disabled, Fadein, Enabled)
they cycle through in time. They only emit trap_triggered when the
player collides with them while they are Enabled. The times the trap
spends in each state, as well as the initial delay are configurable
via the exported time_* variables to allow them being set in the editor.

There are "a few" traps added to both our test levels.