Introduce Game singleton
For now it contains only list of all levels, with their relevant scenes,
to-be-entrance nodes in world and names. It's better to have this kind
of information stored globally since it's useful for more scenes. Eg.
level name is shown at the beginning of given level and also will be
show in world HUD when player steps on its entrance. When Level holds its
name it'd be really cumbersome to load the whole level when player steps
on its entrance in main world just to get that name.