Refactor Game.current_level to hold id instead of dictionary
My through process behind this refactoring is as follows:

I'm going to work on saving the game progress. The structure I'll use
to hold level progress and number of coins picked up will have level id
as its key. It was currently impossible to get id of current level which
made me think that this might be a problem later on regarding other
functionalities as well. The problem was that current_level did hold
level dictionary which didn't include level id (as it was used as its
key). So I had couple of choices:

* To add id to level dictionary and have a duplicity since the id is
  also level key in dictionary of levels.
* To change dictionary of levels to list and use id as key in level
  dictionary, but that would mean going through the list every time we
  needed to find some level.
* To get level id from level scene's property. This would however mean
  that it's gonna be hardly accessible from outside of level.
* To hold level id in Game.current_level and get the particular level
  dictionary via some helper function whenever it's needed.

I found the last choice to be the most sound one since it means keeping
the dictionary of levels (for O(1) access) and moving just level id
around the application whenever it's needed (instead of dictionary).