Add in-game menu to levels and world
For simplicity, it's always present in the level in a hidden state. It
listens to "ui_cancel" action which invokes it and emits signal to
connected nodes. My first choice was to control menu from level / world
but this turned out to be problematic because of pause mechanic and
propagation of input events. It'd be doable, but I found it to be too
complicated for now. Menu doesn't know anything about the pause
mechanic. Pausing is responsibility of level / world in response to
`menu_activated` signal.

There are 2 menus inheriting from `GameMenu`:

* `LevelMenu` which gives the player an option to exit to world
* `WorldMenu` which gives the player an option to exit the game
  completely