Turn off collisions on platform's TileMap
It messed up with movement of player. TileMap's and KinematicBody2D's
collision shapes overlapped. TileMap acted as a static body and it
caused player to (sometimes) not follow movement of the kinematic body.
Unfortunatelly, we cannot use TileMap in kinematic body mode as it lacks
sync with physics which was added for kinematic bodies in 3.1.