Save state of the level scenes when player reaches checkpoint
Enemies and coins have new `serialize` and `deserialize` methods that
take care of saving and loading of relevant properties. When player
reaches checkpoint, `deserialize` is called on all relevant scenes and
state of level in the latest checkpoint is saved. When player dies,
they're moved to spawner of the latest checkpoint, all level scenes
are removed and new loaded scenes are spawned instead. `deserialize` is
called in each of them to make sure they are in proper state.